Connected journal service to firestore
Journals do not save with an id, since id is created by firestore as document name. So when reading from firestore, merge this id into the journal to be able to then save it back again.
Newly created journal entries that are not yet saved do not have an id, so instead use the create method.