Make the selected task a different colour
To do this, I needed all journals to have an id, to identify the selected one. Achieved this by creating an incrementing id for unsaved journals, which get replaced with their firestore id when they are saved.

After saving new or updating existing journals, performed some fancy filtering to keep everything in sync with the database. Ideally I could just observe the collection, and locally store unsaved journals.. but when trying this I always seemed to observe firestore twice, and doubled my unsaved journals. This approach works... for now....

Performed sanitisation on data before writing to firestore. Would love to know a better way to enforce this with typing... but it eludes me for now.