Adding authentication and protected routes
Created an auth service to sign in with google.
Merge user into firestore on login.
Store user in local store, which is how we can know if we already have a user logged in.
Protected routes using both an AuthGuard (check if logged in) or NegateAuthGuard (check if user is already logged in).
Added sign out button on journals page.